Cleanup plugins and simplify server mocks
