+++ src/main/kotlin/no/nav/syfo/application/api/ApiPlugins.kt
@@ -1,78 +1,20 @@
-package no.nav.syfo.application.api.authentication
+package no.nav.syfo.application.api
 
-import com.auth0.jwk.JwkProviderBuilder
-import com.auth0.jwt.JWT
 import io.ktor.client.plugins.*
 import io.ktor.http.*
 import io.ktor.serialization.jackson.*
 import io.ktor.server.application.*
-import io.ktor.server.auth.*
-import io.ktor.server.auth.jwt.*
 import io.ktor.server.metrics.micrometer.*
 import io.ktor.server.plugins.contentnegotiation.*
 import io.ktor.server.plugins.statuspages.*
 import io.ktor.server.response.*
 import io.micrometer.core.instrument.distribution.DistributionStatisticConfig
-import net.logstash.logback.argument.StructuredArguments
 import no.nav.syfo.behandler.api.access.ForbiddenAccessVeilederException
 import no.nav.syfo.behandler.api.person.access.ForbiddenPersonAPIConsumer
-import no.nav.syfo.domain.PersonIdentNumber
 import no.nav.syfo.metric.METRICS_REGISTRY
 import no.nav.syfo.util.configure
 import no.nav.syfo.util.getCallId
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
-import java.net.URL
 import java.time.Duration
-import java.util.concurrent.TimeUnit
-
-private val log: Logger = LoggerFactory.getLogger("no.nav.syfo.application.api.authentication")
-
-fun Application.installJwtAuthentication(
-    jwtIssuerList: List<JwtIssuer>,
-) {
-    install(Authentication) {
-        jwtIssuerList.forEach { jwtIssuer ->
-            configureJwt(
-                jwtIssuer = jwtIssuer,
-            )
-        }
-    }
-}
-
-fun AuthenticationConfig.configureJwt(
-    jwtIssuer: JwtIssuer,
-) {
-    val jwkProviderSelvbetjening = JwkProviderBuilder(URL(jwtIssuer.wellKnown.jwks_uri))
-        .cached(10, 24, TimeUnit.HOURS)
-        .rateLimited(10, 1, TimeUnit.MINUTES)
-        .build()
-    jwt(name = jwtIssuer.jwtIssuerType.name) {
-        verifier(jwkProviderSelvbetjening, jwtIssuer.wellKnown.issuer)
-        validate { credential ->
-            if (hasExpectedAudience(credential, jwtIssuer.acceptedAudienceList)) {
-                JWTPrincipal(credential.payload)
-            } else {
-                log.warn(
-                    "Auth: Unexpected audience for jwt {}, {}",
-                    StructuredArguments.keyValue("issuer", credential.payload.issuer),
-                    StructuredArguments.keyValue("audience", credential.payload.audience)
-                )
-                null
-            }
-        }
-    }
-}
-
-fun hasExpectedAudience(credentials: JWTCredential, expectedAudience: List<String>): Boolean {
-    return expectedAudience.any { credentials.payload.audience.contains(it) }
-}
-
-fun ApplicationCall.personIdent(): PersonIdentNumber? {
-    val token = this.request.headers[HttpHeaders.Authorization]?.removePrefix("Bearer ")
-    val decodedJWT = JWT.decode(token)
-    return decodedJWT.claims["pid"]?.asString()?.let { PersonIdentNumber(it) }
-}
 
 fun Application.installMetrics() {
     install(MicrometerMetrics) {
