+++ src/main/kotlin/no/nav/syfo/application/api/authentication/ApiAuthenticationPlugin.kt
@@ -1,0 +1,53 @@
+package no.nav.syfo.application.api.authentication
+
+import com.auth0.jwk.JwkProviderBuilder
+import io.ktor.server.application.*
+import io.ktor.server.auth.*
+import io.ktor.server.auth.jwt.*
+import net.logstash.logback.argument.StructuredArguments
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
+import java.net.URL
+import java.util.concurrent.TimeUnit
+
+private val log: Logger = LoggerFactory.getLogger("no.nav.syfo.application.api.authentication")
+
+fun Application.installJwtAuthentication(
+    jwtIssuerList: List<JwtIssuer>,
+) {
+    install(Authentication) {
+        jwtIssuerList.forEach { jwtIssuer ->
+            configureJwt(
+                jwtIssuer = jwtIssuer,
+            )
+        }
+    }
+}
+
+fun AuthenticationConfig.configureJwt(
+    jwtIssuer: JwtIssuer,
+) {
+    val jwkProviderSelvbetjening = JwkProviderBuilder(URL(jwtIssuer.wellKnown.jwks_uri))
+        .cached(10, 24, TimeUnit.HOURS)
+        .rateLimited(10, 1, TimeUnit.MINUTES)
+        .build()
+    jwt(name = jwtIssuer.jwtIssuerType.name) {
+        verifier(jwkProviderSelvbetjening, jwtIssuer.wellKnown.issuer)
+        validate { credential ->
+            if (hasExpectedAudience(credential, jwtIssuer.acceptedAudienceList)) {
+                JWTPrincipal(credential.payload)
+            } else {
+                log.warn(
+                    "Auth: Unexpected audience for jwt {}, {}",
+                    StructuredArguments.keyValue("issuer", credential.payload.issuer),
+                    StructuredArguments.keyValue("audience", credential.payload.audience)
+                )
+                null
+            }
+        }
+    }
+}
+
+fun hasExpectedAudience(credentials: JWTCredential, expectedAudience: List<String>): Boolean {
+    return expectedAudience.any { credentials.payload.audience.contains(it) }
+}
