+++ src/main/kotlin/no/nav/syfo/util/RequestUtil.kt
@@ -1,9 +1,11 @@
 package no.nav.syfo.util
 
+import com.auth0.jwt.JWT
 import io.ktor.server.application.*
 import io.ktor.http.*
 import io.ktor.util.pipeline.*
 import net.logstash.logback.argument.StructuredArguments
+import no.nav.syfo.domain.PersonIdentNumber
 
 const val NAV_CALL_ID_HEADER = "Nav-Call-Id"
 const val NAV_PERSONIDENT_HEADER = "nav-personident"
@@ -23,6 +25,12 @@ fun PipelineContext<out Unit, ApplicationCall>.getBearerHeader(): String? {
 
 fun PipelineContext<out Unit, ApplicationCall>.getPersonIdentHeader(): String? {
     return this.call.request.headers[NAV_PERSONIDENT_HEADER]
+}
+
+fun ApplicationCall.personIdentFromToken(): PersonIdentNumber? {
+    val token = this.request.headers[HttpHeaders.Authorization]?.removePrefix("Bearer ")
+    val decodedJWT = JWT.decode(token)
+    return decodedJWT.claims["pid"]?.asString()?.let { PersonIdentNumber(it) }
 }
 
 fun bearerHeader(token: String) = "Bearer $token"
