+++ src/test/kotlin/no/nav/syfo/testhelper/ExternalMockEnvironment.kt
@@ -3,12 +3,13 @@ package no.nav.syfo.testhelper
 import io.ktor.server.netty.*
 import no.nav.common.KafkaEnvironment
 import no.nav.syfo.application.ApplicationState
+import no.nav.syfo.application.Environment
 import no.nav.syfo.testhelper.mocks.*
 
 class ExternalMockEnvironment private constructor() {
     val applicationState: ApplicationState = testAppState()
     val database = TestDatabase()
-    val embeddedEnvironment: KafkaEnvironment = testKafka()
+    private val embeddedEnvironment: KafkaEnvironment = testKafka()
     private val azureAdMock = AzureAdMock()
 
     private val fastlegeRestMock = FastlegeRestMock()
@@ -24,14 +25,16 @@ class ExternalMockEnvironment private constructor() {
         pdlMock.name to pdlMock.server,
     )
 
-    val environment = testEnvironment(
-        azureOpenidConfigTokenEndpoint = azureAdMock.url,
-        kafkaBootstrapServers = embeddedEnvironment.brokersURL,
-        fastlegeRestUrl = fastlegeRestMock.url,
-        syfoPartnerinfoUrl = syfopartnerInfoMock.url,
-        syfoTilgangskontrollUrl = syfoTilgangskontrollMock.url,
-        pdlUrl = pdlMock.url,
-    )
+    val environment: Environment by lazy {
+        testEnvironment(
+            azureOpenidConfigTokenEndpoint = azureAdMock.url(),
+            kafkaBootstrapServers = embeddedEnvironment.brokersURL,
+            fastlegeRestUrl = fastlegeRestMock.url(),
+            syfoPartnerinfoUrl = syfopartnerInfoMock.url(),
+            syfoTilgangskontrollUrl = syfoTilgangskontrollMock.url(),
+            pdlUrl = pdlMock.url(),
+        )
+    }
     val wellKnownInternalAzureAD = wellKnownInternalAzureAD()
     val wellKnownInternalIdportenTokenX = wellKnownInternalIdportenTokenX()
